Split existing image job response helpers
