+++ frontend/src/server/images/execute-image-generation.ts
@@ -28,7 +28,7 @@ import {
   uploadImageToStorage,
 } from '@/server/storage';
 import { createImageThumbnailBatch } from '@/server/image-thumbnails';
-import { buildStoredImageRenderEntries, parseStoredImageRenders, resolveHeroThumbFromRenders } from '@/lib/image-renders';
+import { buildStoredImageRenderEntries, resolveHeroThumbFromRenders } from '@/lib/image-renders';
 import { ensureReusableAsset, upsertLegacyJobOutputs } from '@/server/media-library';
 import {
   formatSupportedImageFormatsLabel,
@@ -55,6 +55,9 @@ import {
 } from '@/lib/image/gptImage2';
 import { computeBillingProductSnapshot } from '@/lib/billing-products';
 import type { BillingProductKey, JobSurface } from '@/types/billing';
+import { buildResponseFromExistingJob, type ExistingImageJobRow } from './existing-image-job-response';
+
+export { buildResponseFromExistingJob } from './existing-image-job-response';
 
 const DISPLAY_CURRENCY = 'USD';
 const DISPLAY_CURRENCY_LOWER: Currency = 'usd';
@@ -78,25 +81,6 @@ type PendingReceipt = {
   vendorAccountId: string | null;
   stripePaymentIntentId?: string | null;
   stripeChargeId?: string | null;
-};
-
-export type ExistingImageJobRow = {
-  job_id: string;
-  user_id: string | null;
-  status: string;
-  progress: number;
-  provider_job_id: string | null;
-  thumb_url: string | null;
-  aspect_ratio: string | null;
-  pricing_snapshot: PricingSnapshot | null;
-  currency: string | null;
-  payment_status: string | null;
-  engine_id: string;
-  engine_label: string;
-  render_ids: unknown;
-  hero_render_id: string | null;
-  message: string | null;
-  settings_snapshot: unknown;
 };
 
 type ExistingImageChargeRow = {
@@ -619,69 +603,6 @@ function fail(
   extras?: Partial<ImageGenerationResponse>
 ): never {
   throw new ImageGenerationExecutionError(message, { mode, code, status, detail, extras });
-}
-
-export function parseResolutionFromSettingsSnapshot(snapshot: unknown): string | null {
-  if (!snapshot || typeof snapshot !== 'object' || Array.isArray(snapshot)) return null;
-  const core = (snapshot as Record<string, unknown>).core;
-  if (!core || typeof core !== 'object' || Array.isArray(core)) return null;
-  const resolution = (core as Record<string, unknown>).resolution;
-  return typeof resolution === 'string' && resolution.trim().length ? resolution : null;
-}
-
-function buildImagesFromExistingJob(job: ExistingImageJobRow): GeneratedImage[] {
-  const parsed = parseStoredImageRenders(job.render_ids);
-  if (parsed.entries.length) {
-    return parsed.entries.map((entry) => ({
-      url: entry.url,
-      thumbUrl: entry.thumbUrl ?? entry.url,
-      width: entry.width ?? null,
-      height: entry.height ?? null,
-      mimeType: entry.mimeType ?? null,
-    }));
-  }
-
-  const heroUrl = job.hero_render_id ? normalizeMediaUrl(job.hero_render_id) ?? job.hero_render_id : null;
-  const thumbUrl = job.thumb_url ? normalizeMediaUrl(job.thumb_url) ?? job.thumb_url : null;
-  if (!heroUrl) return [];
-
-  return [
-    {
-      url: heroUrl,
-      thumbUrl: thumbUrl ?? heroUrl,
-    },
-  ];
-}
-
-export function buildResponseFromExistingJob(args: {
-  job: ExistingImageJobRow;
-  mode: ImageGenerationMode;
-  engineId: string;
-  engineLabel: string;
-  pricing: PricingSnapshot;
-  resolvedAspectRatio: string | null;
-  resolution: string;
-}): ImageGenerationResponse {
-  const images = buildImagesFromExistingJob(args.job);
-  const thumbUrl = args.job.thumb_url ? normalizeMediaUrl(args.job.thumb_url) ?? args.job.thumb_url : null;
-  const pricing = args.job.pricing_snapshot ?? args.pricing;
-
-  return {
-    ok: true,
-    mode: args.mode,
-    images,
-    description: args.job.message ?? null,
-    jobId: args.job.job_id,
-    requestId: args.job.provider_job_id ?? undefined,
-    providerJobId: args.job.provider_job_id ?? undefined,
-    engineId: args.job.engine_id || args.engineId,
-    engineLabel: args.job.engine_label || args.engineLabel,
-    pricing,
-    paymentStatus: args.job.payment_status ?? 'paid_wallet',
-    thumbUrl,
-    aspectRatio: args.job.aspect_ratio ?? args.resolvedAspectRatio,
-    resolution: parseResolutionFromSettingsSnapshot(args.job.settings_snapshot) ?? args.resolution,
-  };
 }
 
 async function insertProvisionalImageJob(executor: QueryExecutor, params: ProvisionalImageJobInsert) {
